Validation logic moved to Session Service + Relevant Tests + code Refactoring
