+++ backend/api-service/src/main/java/com/microtimemanagement/apiservice/service/SessionService.java
@@ -1,6 +1,7 @@
 package com.microtimemanagement.apiservice.service;
 
 import com.microtimemanagement.apiservice.dto.entity.SessionDTO;
+import com.microtimemanagement.apiservice.dto.entity.ValidSessionDTO;
 import com.microtimemanagement.apiservice.model.Session;
 import com.microtimemanagement.apiservice.model.User;
 
@@ -10,7 +11,7 @@ public interface SessionService {
 
     void destroySession(String token);
 
-    void validateSession(Session session);
+    ValidSessionDTO validateSessionForAccessToken(String token);
 
     SessionDTO refreshSession(String token);
 }
